refactor(auth): move the token validation logic to a separate hook
Refs: #37
